Reverted recent changes to item_new_by_level() as they caused segmentation faults.
git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@956 561b6253-61b6-49b5-99ee-40376865e0bd

+++ inc/utils.h
@@ -45,7 +45,7 @@ typedef struct _message_log
 #define rand_1n(n)      (((n) <= 1) ? 1 : g_random_int_range(1,(n)))
 #define rand_0n(n)      (((n) <= 0) ? 0 : g_random_int_range(0,(n)))
 /* need to cast m to gint32 as g_random_int_range returns gint32 */
-#define rand_m_n(m,n)   ((m) == (n) ? (gint32)(m) : g_random_int_range((m),(n)))
+#define rand_m_n(m,n)   ((gint32)(m) == (gint32)(n) ? (gint32)(m) : g_random_int_range((m),(n)))
 #define chance(percent) ((gint32)(percent) >= rand_1n(101))
 
 /* windef.h defines these */
